#16 [REFACTOR] MVVM으로 버튼 이벤트 동작하도록 변경
