+++ presentation/src/main/java/com/tgyuu/soccerfriends/feature/home/HomeFragment.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.viewModels
 import com.tgyuu.soccerfriends.common.base.BaseFragment
-import com.tgyuu.soccerfriends.common.base.repeatOnStarted
 import com.tgyuu.soccerfriends.databinding.FragmentHomeBinding
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -18,16 +17,8 @@ class HomeFragment :
 
         binding.apply {
             viewModel = fragmentViewModel.apply {
-                repeatOnStarted {
-                    eventFlow.collect { handleEvent(it) }
-                }
-            }
-        }
-    }
 
-    private fun handleEvent(event : HomeViewModel.HomeEvent){
-        when(event){
-            is open
+            }
         }
     }
 }
