+++ presentation/src/main/java/com/tgyuu/soccerfriends/feature/scoreboard/ScoreBoardFragment.kt
@@ -5,6 +5,7 @@ import android.view.View
 import androidx.fragment.app.viewModels
 import com.tgyuu.soccerfriends.R
 import com.tgyuu.soccerfriends.common.base.BaseFragment
+import com.tgyuu.soccerfriends.common.base.repeatOnStarted
 import com.tgyuu.soccerfriends.databinding.FragmentScoreBoardBinding
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -19,14 +20,19 @@ class ScoreBoardFragment :
         setStatusBarAndIconColor(R.color.main, StatusBarIconColor.WHITE)
 
         binding.apply {
-
-            scoreBoardBTN.setOnClickListener{
-                if(expandableLayout.isExpanded) expandableLayout.collapse()
-                else expandableLayout.expand()
-            }
-
             viewModel = fragmentViewModel.apply {
+                repeatOnStarted {
+                    eventFlow.collect { handleEvent(it) }
+                }
+            }
+        }
+    }
 
+    private fun handleEvent(event: ScoreBoardViewModel.ScoreBoardEvent) {
+        when (event) {
+            ScoreBoardViewModel.ScoreBoardEvent.ClickButton -> binding.apply {
+                if (expandableLayout.isExpanded) expandableLayout.collapse()
+                else expandableLayout.expand()
             }
         }
     }
