+++ presentation/src/main/res/layout/fragment_score_board.xml
@@ -452,6 +452,7 @@
                 android:textColor="@color/white"
                 android:textSize="15sp"
                 android:textStyle="bold"
+                android:onClick="@{()->viewModel.clickButton()}"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent" />
